Initial commit for unwatch

+++ flaw.js
@@ -1,6 +1,8 @@
 var Discord = require("discord.js");
 var config = require("./config.json");
 var fs = require("fs");
+var wfa = require("write-file-atomic").sync;
+var watchlist = {}
 var bot = new Discord.Client();
 const prefix = "~~~";
 
@@ -22,31 +24,45 @@ bot.on("message", msg => {
     let fromBotChannel = msg.channel.id == config.botTestID;
 
     if (command === "watch") {
-        console.log(`Is someone mentioned" ${msg.mentions.users.size!=0}`)
+        console.log("Watch command");
+        console.log(`Is someone mentioned" ${msg.mentions.users.size!=0}`);
         console.log(messageContent[1]);
         let user = msg.mentions.users.size!=0 ? msg.mentions.users.first() : messageContent[1];
-        msg.delete();
+        if (msg.channel.id != config.botTestID) msg.delete();
         watch(msg.mentions.users.size!=0 ? user.username : user);
-        console.log(`Watching ${msg.mentions.users.size!=0 ? user.username : user}`);
-        msg.author.sendMessage(`Watching ${msg.mentions.users.size!=0 ? user.username : user}`);
+        bot.channels.get(config.botTestID).sendMessage(`Watching ${msg.mentions.users.size!=0 ? user.username : user}`)
+        .then(message => {
+            console.log(`Sent watch message: ${message.content}`);
+            message.delete();
+        })
+        .catch(console.error);
+    }
+
+    if (command === "unwatch") {
+        let user = msg.mentions.users.size!=0 ? msg.mentions.users.first() : messageContent[1];
+        if (msg.channel.id != config.botTestID) msg.delete();
+        unwatch(msg.mentions.users.size!=0 ? user.username : user);
+        bot.channels.get(config.botTestID).sendMessage(`No longer watching ${msg.mentions.users.size!=0 ? user.username : user}`)
+        .then(message => {
+            console.log(`Sent unwatch message: ${message.content}`);
+            message.delete();
+        })
+        .catch(console.error);
     }
 
     if (command === "testDM") {
-        // bot.fetchUser(config.flawBotID);
-        // bot.channels.get(config.botTestID).sendMessage("This is a test DM");
-        // msg.author.sendMessage("This is a test DM");
-        console.log(`Sending test message to: ${bot.users.get(config.flawBotID).username}`);
+        bot.channels.get(config.botTestID).sendMessage("This is a test DM");
+        console.log(`Sending test message`);
     }
 });
 
 bot.on("presenceUpdate", (oldUser, newUser) => {
     try {
-        let users = JSON.parse(fs.readFileSync("./watchlist.json"));
-    if (users[oldUser.user.username]) {
-        if ((oldUser.presence.status === "offline") && (newUser.presence.status != "offline")) {
-            bot.channels.get(config.botTestID).sendMessage(`${oldUser.user.username} is ${newUser.presence.status}`);
+        if (watchlist[oldUser.user.username]) {
+            if ((oldUser.presence.status === "offline") && (newUser.presence.status != "offline")) {
+                bot.channels.get(config.botTestID).sendMessage(`${oldUser.user.username} is ${newUser.presence.status}`);
+            }
         }
-    }
     } catch (error) {
         console.error(error);
     }
@@ -56,10 +72,19 @@ bot.login(config.userToken);
 
 function watch(user) {
     try {
-        let userObject = JSON.parse(fs.readFileSync("./watchlist.json"));
-        userObject[user] = user;
-        console.log(JSON.stringify(user));
-        fs.writeFileSync("./watchlist.json", JSON.stringify(userObject));
+        watchlist[user] = user;
+        console.log(JSON.stringify(watchlist));
+        return;
+    } catch (error) {
+        console.error(error);
+    }
+}
+
+function unwatch(user) {
+    try {
+        delete watchlist[user];
+        console.log(JSON.stringify(watchlist));
+        return;
     } catch (error) {
         console.error(error);
     }
